pyproject.toml
[tool.poetry]
name = "allegro-orders-backup"
version = "1.0.0"
description = "Микросервис для резервного копирования заказов Allegro через API"
authors = ["System Architect <architect@company.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
sqlmodel = "^0.0.14"
pydantic = "^2.5.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
celery = "^5.3.4"
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"
python-multipart = "^0.0.6"
fastapi-users = "^12.1.2"
fastapi-pagination = "^0.12.12"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.10.1"
pytz = "^2025.2"
sqlalchemy-celery-beat = "^0.8.4"
requests = "^2.32.4"
flower = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 